Fix auth issues in postsController
- Extend request object with decoded token in secured routes.

+++ routes/postsController.ts
@@ -1,22 +1,50 @@
 import { Router } from "express";
-import { secureRouteForPosts } from "../utils/secureRouteForPosts";
+import { extractAuthToken } from "../utils/extractAuthToken";
 import { Post } from "../models/post";
 import { body, validationResult } from "express-validator";
+import { checkIsValidRole } from "../utils/checkIsValidRole";
 
 const postsController = Router();
 
-postsController.get("/", async (req, res) => {
-  try {
-    const posts = await Post.find().populate("author", "_id username");
-    return res.json({ posts });
-  } catch (error) {
-    return res.status(500).json({ error: "Error on loading posts" });
+postsController.get(
+  "/",
+  async (req, res, next) => {
+    const { type } = req.query;
+
+    if (type === "own") {
+      return next();
+    }
+
+    try {
+      const posts = await Post.find(
+        { isPublished: true },
+        { isPublished: false }
+      )
+        .sort({ createdAt: -1 })
+        .populate("author", "_id username");
+      return res.json({ posts });
+    } catch (error) {
+      return res.status(500).json({ error: "Error on loading posts" });
+    }
+  },
+  extractAuthToken,
+  checkIsValidRole,
+  async (req, res) => {
+    try {
+      const posts = await Post.find({ author: req!.user!.id }).sort({
+        createdAt: -1,
+      });
+      return res.status(200).json({ posts });
+    } catch (error) {
+      return res.status(500).json({ error: "Unknown error has occured" });
+    }
   }
-});
+);
 
 postsController.post(
   "/",
-  secureRouteForPosts,
+  extractAuthToken,
+  checkIsValidRole,
   body("title")
     .trim()
     .escape()
@@ -25,7 +53,6 @@ postsController.post(
     .withMessage("Title can't be empty"),
   body("description").trim().escape(),
   body("content").trim().escape(),
-  body("author").trim().escape(),
   async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -36,7 +63,7 @@ postsController.post(
       title: req.body.title,
       description: req.body.description,
       content: req.body.content,
-      author: req.body.author,
+      author: req!.user!.id,
     });
     try {
       await post.save();
@@ -54,7 +81,7 @@ postsController.get("/:postId", async (req, res) => {
   try {
     const post = await Post.findById(postId).populate("author", "_id username");
     if (!post) return res.status(404).json({ error: "Post not found" });
-    return res.json(post);
+    return res.json({ post });
   } catch (error) {
     return res.status(500).json({ error: "Error on loading post" });
   }
@@ -62,7 +89,8 @@ postsController.get("/:postId", async (req, res) => {
 
 postsController.put(
   "/:postId",
-  secureRouteForPosts,
+  extractAuthToken,
+  checkIsValidRole,
   body("isPublished").isBoolean().optional(),
   body("title")
     .trim()
@@ -109,11 +137,23 @@ postsController.put(
       }
     });
     if (Object.keys(updatePayload).length === 0) {
-      return res.status(200).json({ message: "Empty payload" });
-    }
-    try {
-      await Post.findByIdAndUpdate(req.params.postId, updatePayload);
       return res.status(200).json({ message: "Post update was successfull" });
+    }
+
+    const { postId } = req.params;
+
+    try {
+      const post = await Post.findById(postId);
+      if (!post) {
+        return res.status(404).json({ error: "There is no such post" });
+      }
+
+      if (post.author.toString() === req!.user!.id) {
+        await Post.findByIdAndUpdate(postId, updatePayload);
+        return res.status(200).json({ message: "Post update was successfull" });
+      }
+
+      return res.status(403).json({ error: "You have no permission for this" });
     } catch (error) {
       return res
         .status(500)
@@ -122,14 +162,31 @@ postsController.put(
   }
 );
 
-postsController.delete("/:postId", secureRouteForPosts, async (req, res) => {
-  const postId = req.params;
-  try {
-    const post = await Post.findByIdAndDelete(postId);
-    res.status(200).json({ message: "Post has been successfully deleted" });
-  } catch (error) {
-    return res.status(500).json({ error: "Error on deleting post" });
+postsController.delete(
+  "/:postId",
+  extractAuthToken,
+  checkIsValidRole,
+  async (req, res) => {
+    const { postId } = req.params;
+    try {
+      const post = await Post.findById(postId);
+
+      if (!post) return res.status(204).json({});
+      if (
+        req!.user!.id === post?.author.toString() ||
+        (req!.user!.role === "admin" && post?.isPublished === true)
+      ) {
+        await Post.findByIdAndDelete(postId);
+        return res.status(204).json({});
+      }
+
+      return res
+        .status(403)
+        .json({ error: "You haven't got permission for this" });
+    } catch (error) {
+      return res.status(500).json({ error: "Error on deleting post" });
+    }
   }
-});
+);
 
 export default postsController;
